Move the blocked-agent list to Security
The just-in-time processing gate runs whether or not feed auth is
enabled, so filing the blocked-agent list under Authenticated Feeds
implied a dependency that does not exist. Placement only, no
behaviour change.
