+++ docs/configuration.md
@@ -192,7 +192,7 @@ API: `titleSkipPatterns` (array of strings, max 50 patterns, 200 characters each
 
 ### Blocked user agents for just-in-time processing
 
-Settings > Authenticated Feeds has an **Agents that skip processing** list, empty by default. If a listed User-Agent asks for an episode MinusPod has not processed yet, MinusPod answers with a 302 to the original audio URL rather than queueing a transcription and ad-detection run. Already-processed episodes are unaffected and still serve the cut version to every client, listed or not.
+Settings > Security has an **Agents that skip processing** list, empty by default. If a listed User-Agent asks for an episode MinusPod has not processed yet, MinusPod answers with a 302 to the original audio URL rather than queueing a transcription and ad-detection run. Already-processed episodes are unaffected and still serve the cut version to every client, listed or not.
 
 Matching is case-insensitive, and a bare pattern matches anywhere in the agent string. Start a pattern with `^` to anchor it to the beginning, which short strings like `atc/` need so they cannot match in the middle of an unrelated agent. Each entry is limited to 200 characters; blank or whitespace-only entries are dropped when saved. The [opawg/user-agents](https://github.com/opawg/user-agents) registry lists the real strings crawlers and podcast apps send, useful when picking a pattern.
